cleanup function placements
