+++ utils/constants.py
@@ -1,0 +1,5 @@
+import os
+
+SCRIPT_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+
+OUTPUT_BASE_DIR = os.path.join(SCRIPT_DIR, 'data')
